refactor: 리스너를 viewModel에서 구현하도록 변경
